Chrome Extension Backend/requirements.txt
flask
gunicorn
Flask-Failsafe
requests
python-dotenv